Fix and improve the scripts, plus renamed prod-test-manager to run-intg-test
